feat: Add support for MPD Tiles of thumbnail images in <dash-video/> element (#177)
Added support for tiles of thumbnail images provided via MPD as specified in Guidelines for Implementation:
DASH-IF Interoperability Points section "6.2.6. Tiles of thumbnail images".

The changes include checking for image representations, creating a WebVTT Blob file, and creating the thumbnail <track/> element inside <dash-video/> with it's source pointing to the file. This is done only on stream initialization. We don't update an existing thumbnail <track/> element so we don't overwrite whatever is set on the HTML.